refactor(runner): extract runtime-contracts.ts from sandbox_agent.ts
Move the engine's shared interfaces and types (SandboxAgentDeps, CurrentTurn,
ParkedApproval, ResumeApprovalInput, RunTurnOptions, sendLastMessageOnly,
SessionEnvironment, AcquireEnvironmentResult, RUN_LIMIT_TRIPPED) into
runtime-contracts.ts. Monolith re-exports the public ones so external importers
are unchanged. Pure code move; typecheck clean; 1158 runner unit tests pass.
