+++ services/runner/src/engines/sandbox_agent.ts
@@ -54,7 +54,6 @@ import {
   extractApprovalDecisions,
   extractClientToolOutputs,
   type ClientToolOutcome,
-  type Responder,
 } from "../responder.ts";
 import type { ClientToolRelay } from "../tools/client-tool-relay.ts";
 import {
@@ -65,7 +64,6 @@ import {
   type AgentRunRequest,
   type AgentRunResult,
   type EmitEvent,
-  type HarnessCapabilities,
   type ToolCallbackContext,
   resolvePromptText,
   resolveRunSessionId,
@@ -112,10 +110,7 @@ import {
   PendingApprovalLatch,
   permissionsFromRequest,
 } from "../permission-plan.ts";
-import {
-  attachPermissionResponder,
-  type ParkedApprovalGateType,
-} from "./sandbox_agent/acp-interactions.ts";
+import { attachPermissionResponder } from "./sandbox_agent/acp-interactions.ts";
 import {
   PAUSED,
   PendingApprovalPauseController,
@@ -136,11 +131,7 @@ import {
 import { buildSandboxProvider } from "./sandbox_agent/provider.ts";
 import { loadRunnerConfig } from "../config/runner-config.ts";
 import { DaytonaReconnectTerminalError } from "./sandbox_agent/daytona-provider.ts";
-import {
-  buildRunPlan,
-  type BuildRunPlanDeps,
-  type RunPlan,
-} from "./sandbox_agent/run-plan.ts";
+import { buildRunPlan } from "./sandbox_agent/run-plan.ts";
 import { priorMessages } from "./sandbox_agent/transcript.ts";
 import { resolveRunUsage } from "./sandbox_agent/usage.ts";
 import { prepareWorkspace } from "./sandbox_agent/workspace.ts";
@@ -184,7 +175,6 @@ import {
   eligibleAgentSessionId,
   nextTurnIndex,
   sessionContinuityStore,
-  type SessionContinuityStore,
 } from "./sandbox_agent/session-continuity.ts";
 import {
   projectScopeFor,
@@ -200,12 +190,32 @@ import {
   isTransportEndpointDisconnected,
   containsTransportEndpointDisconnected,
 } from "./sandbox_agent/runtime-policy.ts";
+import {
+  sendLastMessageOnly,
+  RUN_LIMIT_TRIPPED,
+  type SandboxAgentDeps,
+  type CurrentTurn,
+  type ParkedApproval,
+  type ResumeApprovalInput,
+  type RunTurnOptions,
+  type SessionEnvironment,
+  type AcquireEnvironmentResult,
+} from "./sandbox_agent/runtime-contracts.ts";
 
 export {
   buildTurnText,
   messageTranscript,
 } from "./sandbox_agent/transcript.ts";
 export { toAcpMcpServers } from "./sandbox_agent/mcp.ts";
+export {
+  sendLastMessageOnly,
+  type SandboxAgentDeps,
+  type ParkedApproval,
+  type ResumeApprovalInput,
+  type RunTurnOptions,
+  type SessionEnvironment,
+  type AcquireEnvironmentResult,
+} from "./sandbox_agent/runtime-contracts.ts";
 
 function log(message: string): void {
   process.stderr.write(`[sandbox-agent] ${message}\n`);
@@ -284,223 +294,6 @@ export async function destroyInFlightSandboxesForSession(
     new Promise<void>((resolve) => setTimeout(resolve, timeoutMs)),
   ]);
 }
-
-export interface SandboxAgentDeps extends BuildRunPlanDeps {
-  startSandboxAgent?: typeof SandboxAgent.start;
-  createPersist?: () => InMemorySessionPersistDriver;
-  createOtel?: typeof createSandboxAgentOtel;
-  buildDaemonEnv?: typeof buildDaemonEnv;
-  resolveDaemonBinary?: typeof resolveDaemonBinary;
-  buildSandboxProvider?: typeof buildSandboxProvider;
-  createCookieFetch?: typeof createCookieFetch;
-  createAcpFetch?: typeof createAcpFetch;
-  prepareWorkspace?: typeof prepareWorkspace;
-  prepareDaytonaPiAssets?: typeof prepareDaytonaPiAssets;
-  uploadToolMcpAssets?: typeof uploadToolMcpAssets;
-  probeCapabilities?: typeof probeCapabilities;
-  applyModel?: typeof applyModel;
-  startToolRelay?: typeof startToolRelay;
-  localRelayHost?: typeof localRelayHost;
-  sandboxRelayHost?: typeof sandboxRelayHost;
-  signSessionMountCredentials?: typeof signSessionMountCredentials;
-  signAgentMountCredentials?: typeof signAgentMountCredentials;
-  mountStorage?: typeof mountStorage;
-  mountStorageRemote?: typeof mountStorageRemote;
-  unmountStorage?: typeof unmountStorage;
-  discoverTunnelEndpoint?: typeof discoverTunnelEndpoint;
-  /** Per-harness transcript mounts (remote only; see mount.ts). */
-  mountHarnessSessionDirs?: typeof mountHarnessSessionDirs;
-  responderFactory?: (request: AgentRunRequest) => Responder;
-  resolveRunLimits?: typeof resolveRunLimits;
-  createRunLimits?: typeof createRunLimits;
-  /** Session-continuity store override (tests inject their own; default is the process singleton). */
-  sessionContinuityStore?: SessionContinuityStore;
-  /** Durable read-back/write-forward of the continuity store (tests inject fakes). */
-  hydrateHarnessSessionFromDurable?: typeof hydrateHarnessSessionFromDurable;
-  syncHarnessSessionDurable?: typeof syncHarnessSessionDurable;
-  /** Durable read/write of the sandbox pointer, for the remote reconnect ladder. */
-  readStoredSandboxPointer?: typeof readStoredSandboxPointer;
-  clearSandboxPointer?: typeof clearSandboxPointer;
-  writeSandboxPointer?: typeof writeSandboxPointer;
-  /**
-   * Resolve `{replicaId, ownerReplicaId}` for a session-owned local-sandbox run, so
-   * `acquireEnvironment` can fail loudly instead of silently cold-starting on a non-owner
-   * replica. The default claims the `owner` affinity key via the coordination plane and reads
-   * back the actual owner (`claimSessionOwnership`); tests inject their own. `authorization` is
-   * the run credential (the claim authenticates as the invoke caller).
-   */
-  resolveLocalRunnerOwner?: (
-    sessionId: string,
-    authorization: string,
-  ) => Promise<{ replicaId: string; ownerReplicaId: string | undefined }>;
-  log?: Log;
-}
-
-/**
- * Race sentinel: a run-limits deadline (total/idle/TTFB/per-tool-call) tripped mid-turn. Distinct
- * from `PAUSED` so the prompt race can tell a human pause (keep the session) from a wedge deadline
- * (end the turn as an error, letting the caller's teardown reclaim the sandbox).
- */
-const RUN_LIMIT_TRIPPED = Symbol("run-limit-tripped");
-
-/**
- * The per-turn sink the session-lifetime listeners demux into. `runTurn` swaps a fresh one in
- * at turn start (`env.currentTurn`) and the dispatch clears it at turn end. The `sandbox-agent`
- * listener registries are plain Sets — an event with no listener is dropped and a permission
- * request with no listener is CANCELLED — so the listeners stay attached for the session's whole
- * life and route into whichever turn is active, with no detach/attach window between turns.
- */
-interface CurrentTurn {
-  run: ReturnType<typeof createSandboxAgentOtel>;
-  pause: PendingApprovalPauseController;
-  toolRelay?: { ready?: Promise<void>; stop: () => Promise<void> };
-  /** Route a session/update for the active turn (suppress + handleUpdate + pause re-sweep). */
-  handleUpdate: (update: unknown) => void;
-  /** Route a permission reverse-RPC for the active turn (built by attachPermissionResponder). */
-  onPermissionRequest?: (req: unknown) => void;
-}
-
-/**
- * A permission gate that paused the turn and can be answered later on the SAME live session.
- * Recorded for a Claude ACP permission gate (keep-alive slice 2) or a Pi ACP permission gate
- * (Pi approval parking: the gate rides the extension's `ctx.ui.confirm` onto the same ACP
- * permission plane). NOT recorded for a client-tool MCP pause — that cannot be answered across
- * a turn boundary and stays on the cold path. Existence of this record is what makes the
- * dispatch park a paused session in `awaiting_approval` instead of tearing it down.
- */
-export interface ParkedApproval {
-  /** Which gate paused; the dispatch resumes only a recognized type and treats others as cold. */
-  gateType: ParkedApprovalGateType;
-  /** The ACP permission-request id, answered later via `session.respondPermission`. */
-  permissionId: string;
-  /** The gated tool call's id — matched against the incoming approval envelope's toolCallId. */
-  toolCallId: string;
-  /** The gated tool name (logging + the durable interaction row); never its args, in logs. */
-  toolName: string | undefined;
-  /** The gated call's original args, used to seed the resume turn's trace/egress tool span. */
-  args: unknown;
-  /** The durable interaction row token, resolved on the answer via the onResolveInteraction hook. */
-  interactionToken: string;
-  /** The held original `prompt()` promise; the resume awaits it after `respondPermission`. */
-  promptPromise?: Promise<unknown>;
-}
-
-/** Answer a parked Claude ACP permission gate on the live session (the keep-alive resume input). */
-export interface ResumeApprovalInput {
-  permissionId: string;
-  reply: "once" | "reject";
-  toolCallId: string;
-  toolName: string | undefined;
-  args: unknown;
-  interactionToken: string;
-  promptPromise?: Promise<unknown>;
-}
-
-/** Per-turn options for `runTurn`. Absent (flag off / cold) means today's byte-identical path. */
-export interface RunTurnOptions {
-  /** A live continuation: send only the new user text instead of the full cold transcript. */
-  continuation?: boolean;
-  /**
-   * The session was rehydrated via `session/load` (the patched `resumeSession`), so the harness
-   * already holds the prior turns natively. Like `continuation`, the prompt is only the new user
-   * text; `buildTurnText` must not run. Distinct field from `continuation` because the two arrive
-   * through different acquire paths (live pool checkout vs a fresh cold acquire that loaded an
-   * old session) — `runTurn` treats them identically for the text-selection decision.
-   */
-  loaded?: boolean;
-  /**
-   * Keep-alive approval park mode: on a Claude ACP permission gate the pause keeps the session
-   * alive (no settle/abort/destroy) so a later resume can answer it. A non-parkable pause (Pi
-   * relay, client tool) still tears down exactly as today, so this is safe to set on any eligible
-   * keep-alive turn.
-   */
-  approvalParkMode?: boolean;
-  /** A live approval resume: answer the parked gate and stream the continued prompt's events. */
-  resume?: ResumeApprovalInput;
-}
-
-/**
- * Send only the new user text (not the full cold transcript) when the harness already holds the
- * prior turns: a live continuation, or a session rehydrated via `session/load`. `runTurn` calls
- * this, so a test that pins it pins the shipped decision.
- */
-export function sendLastMessageOnly(opts: RunTurnOptions): boolean {
-  return Boolean(opts.continuation || opts.loaded);
-}
-
-/**
- * A session-scoped environment that can serve many turns. Everything expensive to build lives
- * here (sandbox, session, internal tool-MCP server, mounted cwd, relay/temp dirs); `destroy()`
- * is the one complete idempotent teardown the pool, the shutdown handler, and the cold path all
- * call. Per-turn state rides `currentTurn`, swapped in by `runTurn`.
- */
-export interface SessionEnvironment {
-  plan: RunPlan;
-  logger: Log;
-  deps: SandboxAgentDeps;
-  sandbox: any;
-  session: any;
-  sessionId: string;
-  model: string | undefined;
-  capabilities: HarnessCapabilities;
-  strictModel: boolean;
-  toolCallIndex: ReturnType<typeof createToolCallCorrelationIndex>;
-  /** The current turn's client-tool relay, read by the deferred ref baked into the MCP server. */
-  clientToolRelayRef: { current?: ClientToolRelay };
-  mcpAbort: AbortController;
-  runAgentDir: string | undefined;
-  otlpAuthFilePath: string | undefined;
-  mountCreds: MountCredentials | null;
-  agentMountCreds?: MountCredentials | null;
-  /** The mount's owning project id (keep-alive pool key FALLBACK scope, preferred is
-   * `runContext.project.id`); undefined when there is no mount. */
-  mountProjectId?: string;
-  /** This run's resolved project scope (`projectScopeFor`: run-context preferred, mount
-   * fallback) — the same scope `poolKeyFor` keys on. Undefined when neither source yields
-   * one; a scoped `/kill` can then never claim this sandbox (see `destroyInFlightSandboxesForSession`). */
-  projectScopeId?: string;
-  /** This acquire resumed the harness's native session via `session/load` (not cold). */
-  loadedFromContinuity: boolean;
-  /** A remote, session-owned run whose sandbox can be parked (warm) rather than deleted at end. */
-  resumable: boolean;
-  /** The conversation turn index this acquire's continuity record was read/written at. */
-  continuityTurnIndex: number | undefined;
-  // Mutable teardown/turn state shared across acquire, runTurn, and destroy.
-  sessionDestroyRequested: boolean;
-  mountedCwd: string | undefined;
-  agentMountedPath?: string;
-  durableCwdSafeToDelete: boolean;
-  workspace: { cleanup: () => Promise<void> } | undefined;
-  runtimeRemount: Promise<boolean> | undefined;
-  closeToolMcp: (() => Promise<void>) | undefined;
-  currentTurn?: CurrentTurn;
-  /**
-   * The unique ACP tool-call ids the LAST completed turn emitted (reset at each turn start).
-   * The keep-alive dispatch folds them into the expected next-history fingerprint at park time,
-   * so a tool-using turn still matches its own continuation (the FE keeps assistant tool parts).
-   */
-  lastTurnToolCallIds: string[];
-  /**
-   * The Claude ACP permission gate the LAST turn paused on, or undefined. Set only for a harness
-   * ACP permission gate, reset at each turn start; the dispatch reads it after a paused turn to
-   * decide whether to park in `awaiting_approval` and, on the next request, how to resume.
-   */
-  parkedApproval?: ParkedApproval;
-  /**
-   * How many Claude ACP permission gates resolved to pendingApproval THIS turn (reset at turn
-   * start). More than one means parallel gates the single-gate resume cannot answer, so the
-   * dispatch does not park (tears down cold as today).
-   */
-  approvalGateCount: number;
-  destroyed: boolean;
-  /** Complete, idempotent teardown selected from the typed teardown reason. */
-  destroy: (opts?: { reason?: TeardownReason }) => Promise<void>;
-  /** End the active turn: clear the current-turn sink (called before a park). */
-  clearTurn: () => void;
-}
-
-export type AcquireEnvironmentResult =
-  { ok: true; env: SessionEnvironment } | { ok: false; error: string };
 
 /**
  * Sign the session's durable mount up front so keep-alive can build a pool key (the mount's
